fix: compiler errors in all files & get rid of unique_ptr for Bruno

+++ clique-solver/reductions/k_core.cpp
@@ -11,7 +11,7 @@ void apply_k_core(Graph &G, vector<Vertex*>& maximum_clique) {
 
 void apply_k_core(Graph &G, int lowerCliqueBound)
 {
-    while(G.min_degree < lowerCliqueBound - 1)
+    while((int) G.min_degree < lowerCliqueBound - 1)
     {
         Vertex* lowest_degree_vertex = G.deg_lists[G.min_degree][0];
         G.MM_discard_vertex(lowest_degree_vertex);
